+++ build.gradle
@@ -22,7 +22,7 @@ plugins {
 
 project.ext {
   group = 'com.newrelic.instrumentation'
-  javaAgentVersion = '6.4.0'
+  javaAgentVersion = '6.0.0'
 
   // Aligned with minimum Java major version supported by latest Java Agent
   javaVersion = JavaVersion.VERSION_1_8
