--- reactor-core-2.x/src/main/java/reactor/core/dispatch/AbstractLifecycleDispatcher.java
@@ -1,9 +1,0 @@
-package reactor.core.dispatch;
-
-import com.newrelic.api.agent.weaver.MatchType;
-import com.newrelic.api.agent.weaver.Weave;
-
-@Weave(type=MatchType.BaseClass)
-public abstract class AbstractLifecycleDispatcher {
-
-}
