+++ reactor-core-3.1/build.gradle
@@ -4,7 +4,7 @@
 apply plugin: 'java'
 
 dependencies {
-   implementation 'io.projectreactor:reactor-core:3.1.0.RELEASE'
+   implementation 'io.projectreactor:reactor-core:3.2.11.RELEASE'
 
    // New Relic Java Agent dependencies
    implementation 'com.newrelic.agent.java:newrelic-agent:6.4.0'
@@ -19,13 +19,13 @@ dependencies {
 
 jar {
   manifest {
-    attributes 'Implementation-Title': 'com.newrelic.instrumentation.labs.reactor-core-3.1'
-    attributes 'Implementation-Vendor': 'New Relic Labs'
-    attributes 'Implementation-Vendor-Id': 'com.newrelic.labs'
+    attributes 'Implementation-Title': 'com.newrelic.instrumentation.reactor-core-3.1'
+    attributes 'Implementation-Vendor': 'New Relic'
+    attributes 'Implementation-Vendor-Id': 'com.newrelic'
     attributes 'Implementation-Version': 1.0
   }
 }
 
 verifyInstrumentation {
-	passes 'io.projectreactor:reactor-core:[3.1.0.RELEASE,3.3.0.RELEASE)'
+	passes 'io.projectreactor:reactor-core:[3.1.0.RELEASE,)'
 }
