--- reactor-core-3.1/src/main/java/reactor/core/publisher/Flux_instrumentation.java
@@ -1,18 +1,0 @@
-package reactor.core.publisher;
-
-import com.newrelic.api.agent.weaver.MatchType;
-import com.newrelic.api.agent.weaver.Weave;
-import com.newrelic.api.agent.weaver.Weaver;
-import com.nr.instrumentation.reactor.ReactorUtils;
-
-@Weave(originalName="reactor.core.publisher.Flux",type = MatchType.BaseClass)
-public abstract class Flux_instrumentation {
-	
-	protected static <T> Flux<T> onAssembly(Flux<T> source) {
-		if(!ReactorUtils.initialized) {
-			ReactorUtils.initialize();
-		}
-		return Weaver.callOriginal();
-	}
-
-}
