--- reactor-core-3.3/build.gradle
@@ -1,26 +1,0 @@
-
-// Build.gradle generated for instrumentation module reactor-core-3.1
-
-apply plugin: 'java'
-
-dependencies {
-   implementation 'io.projectreactor:reactor-core:3.3.0.RELEASE'
-
-   // New Relic Java Agent dependencies
-   implementation 'com.newrelic.agent.java:newrelic-agent:6.4.0'
-   implementation 'com.newrelic.agent.java:newrelic-api:6.4.0'
-   implementation fileTree(include: ['*.jar'], dir: '../libs')
-}
-
-jar {
-  manifest {
-    attributes 'Implementation-Title': 'com.newrelic.instrumentation.labs.reactor-core-3.3'
-    attributes 'Implementation-Vendor': 'New Relic Labs'
-    attributes 'Implementation-Vendor-Id': 'com.newrelic.labs'
-    attributes 'Implementation-Version': 1.0
-  }
-}
-
-verifyInstrumentation {
-	passes 'io.projectreactor:reactor-core:[3.3.0.RELEASE,)'
-}
