--- reactor-core-3.3/src/main/java/reactor/core/publisher/Mono_instrumentation.java
@@ -1,17 +1,0 @@
-package reactor.core.publisher;
-
-import com.newrelic.api.agent.weaver.MatchType;
-import com.newrelic.api.agent.weaver.Weave;
-import com.newrelic.api.agent.weaver.Weaver;
-import com.nr.instrumentation.reactor.ReactorUtils;
-
-@Weave(originalName="reactor.core.publisher.Mono",type = MatchType.BaseClass)
-public abstract class Mono_instrumentation {
-	
-	protected static <T> Mono<T> onAssembly(Mono<T> source) {
-		if(!ReactorUtils.initialized) {
-			ReactorUtils.initialize();
-		}
-		return Weaver.callOriginal();
-	}
-}
