+++ settings.gradle
@@ -1,5 +1,4 @@
 rootProject.name = 'java-instrumentation-template'
 include 'reactor-core-2.x'
 include 'reactor-core-3.1'
-include 'reactor-core-3.3'
 include 'reactor-finder'
